Add ADCS board as a new target board
This commit adds the second target board, ADCS, for SC-Sat1 TRCH. For
the ADCS board, all hardware configuration is set to FPGA; thus, we
have nothing to do in TRCH.

The change intentionally builds all I/O board source code but links
only with specified board. This is to ensure that all change is
buildable for both boards and removes the artifacts when `make clean`.

The default board board is "main". If you want to build for ADCS, do:

    make TARGET_BOARD=adcs

Make sure to clean first when you change the board. It doesn't detect
the change and build from scratch yet.

The build test for GitHub Actions is also updated to specify
"TARGET_BOARD".

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>